pyproject.toml
[project]
name = "vpp-opt-test-qqder"
version = "0.1.0"
description = "Virtual Power Plant Optimization Scheduling Framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.3.0",
    "cvxpy>=1.4.0",
    "pulp>=2.7.0",
    "matplotlib>=3.8.0",
    "plotly>=5.17.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.25.0",
    "oemof.solph>=0.5.0",
    "pyomo>=6.6.0",
    "PyYAML>=6.0",
    "scipy>=1.11.0",
    "psutil>=5.9.0"
]
